+++ R/netcomparison.R
@@ -519,7 +519,7 @@ print.netcomparison <- function(x,
   
   sm <- sm.lab <- x$x$sm
   ##
-  relative <- is.relative.effect(sm) | sm == "VE"
+  relative <- is_relative_effect(sm) | sm == "VE"
   #
   if (!backtransf & relative)
     sm.lab <- paste0("log", if (sm == "VE") "VR" else sm)