+++ R/subgroup.netmeta.R
@@ -16,6 +16,13 @@
 #'   used to estimate the between-study variance \eqn{\tau^2} and its
 #'   square root \eqn{\tau}. Either \code{"DL"}, \code{"REML"}, or
 #'   \code{"ML"}, can be abbreviated.
+#' @param level.ma The level used to calculate confidence intervals
+#'   for network estimates.
+#' @param backtransf A logical indicating whether results should be
+#'   back transformed in printouts and forest plots. If
+#'   \code{backtransf = TRUE}, results for \code{sm = "OR"} are
+#'   presented as odds ratios rather than log odds ratios, for
+#'   example.
 #' @param nchar.trts A numeric defining the minimum number of
 #'   characters used to create unique treatment names
 #'   (see \code{\link{netmeta}}).
@@ -41,6 +48,8 @@
 #' @param JAMA.pval A logical specifying whether p-values for test of
 #'   overall effect should be printed according to JAMA reporting
 #'   standards.
+#' @param print.se A logical specifying whether standard errors should be
+#'   printed.
 #' @param print.tau2 A logical specifying whether between-study
 #'   variance \eqn{\tau^2} should be printed.
 #' @param print.tau A logical specifying whether \eqn{\tau}, the
@@ -87,6 +96,8 @@
 subgroup.netmeta <- function(x, subgroup, only.connected = FALSE,
                              common = x$common, random = x$random,
                              method.tau = x$method.tau,
+                             level.ma = x$level.ma,
+                             backtransf = x$backtransf,
                              nchar.trts = x$nchar.trts,
                              ...) {
   
@@ -95,6 +106,8 @@ subgroup.netmeta <- function(x, subgroup, only.connected = FALSE,
   chklogical(only.connected)
   chklogical(common)
   chklogical(random)
+  chklevel(level.ma)
+  chklogical(backtransf)
   #
   method.tau <- setchar(method.tau, c("DL", "REML", "ML"))
   #
@@ -110,12 +123,16 @@ subgroup.netmeta <- function(x, subgroup, only.connected = FALSE,
   #
   subgroup <- catch("subgroup", match.call(), x$data, sys.frame(sys.parent()))
   chknull(subgroup)
-  levs <- sort(unique(subgroup))
+  if (is.factor(subgroup))
+    levs <- levels(subgroup)
+  else
+    levs <- sort(unique(subgroup))
   #
   matNA <- x$TE.common[x$seq, x$seq]
   matNA[!is.na(matNA)] <- NA
   #
-  TE.common <- seTE.common <- TE.random <- seTE.random <-
+  TE.common <- seTE.common <- lower.common <- upper.common <-
+    TE.random <- seTE.random <- lower.random <- upper.random <-
     tau2.matrix <- tau.matrix <- A.matrix <-
     array(NA, dim = c(nrow(matNA), ncol(matNA), length(levs)),
           dimnames = list(rownames(matNA), colnames(matNA), levs))
@@ -142,7 +159,8 @@ subgroup.netmeta <- function(x, subgroup, only.connected = FALSE,
       dat.ks <- subset(dat.k, .studlab %in% studies.s)
       #
       net.ks <- netmeta(.TE, .seTE, .treat1, .treat2, .studlab, data = dat.ks,
-                        sm = x$sm, method.tau = x$method.tau)
+                        sm = x$sm, method.tau = method.tau,
+                        level.ma = level.ma, backtransf = backtransf)
       if (nc.k$n.subnets > 1)
         networks[[k]][[s]] <- net.ks
       else
@@ -152,9 +170,13 @@ subgroup.netmeta <- function(x, subgroup, only.connected = FALSE,
         for (j in net.ks$trts) {
           TE.common[i, j, k] <- net.ks$TE.common[i, j]
           seTE.common[i, j, k] <- net.ks$seTE.common[i, j]
+          lower.common[i, j, k] <- net.ks$lower.common[i, j]
+          upper.common[i, j, k] <- net.ks$upper.common[i, j]
           #
           TE.random[i, j, k] <- net.ks$TE.random[i, j]
           seTE.random[i, j, k] <- net.ks$seTE.random[i, j]
+          lower.random[i, j, k] <- net.ks$lower.random[i, j]
+          upper.random[i, j, k] <- net.ks$upper.random[i, j]
           #
           tau2.matrix[i, j, k] <- net.ks$tau2
           tau.matrix[i, j, k] <- net.ks$tau
@@ -166,8 +188,14 @@ subgroup.netmeta <- function(x, subgroup, only.connected = FALSE,
     #
     diag(TE.common[, , k]) <- NA
     diag(seTE.common[, , k]) <- NA
+    diag(lower.common[, , k]) <- NA
+    diag(upper.common[, , k]) <- NA
+    #
     diag(TE.random[, , k]) <- NA
     diag(seTE.random[, , k]) <- NA
+    diag(lower.random[, , k]) <- NA
+    diag(upper.random[, , k]) <- NA
+    #
     diag(tau2.matrix[, , k]) <- NA
     diag(tau.matrix[, , k]) <- NA
     diag(A.matrix[, , k]) <- NA
@@ -186,6 +214,8 @@ subgroup.netmeta <- function(x, subgroup, only.connected = FALSE,
                      k = A.matrix[trts[i], trts[j], k],
                      TE = TE.common[trts[i], trts[j], k],
                      seTE = seTE.common[trts[i], trts[j], k],
+                     lower = lower.common[trts[i], trts[j], k],
+                     upper = upper.common[trts[i], trts[j], k],
                      Q = m.ij$Q, df.Q = m.ij$df.Q, pval.Q = m.ij$pval.Q)
         #
         if (nrow(res.c) == 0)
@@ -200,7 +230,7 @@ subgroup.netmeta <- function(x, subgroup, only.connected = FALSE,
     for (j in (i + 1):length(trts)) {
       m.ij <- metagen(TE.random[trts[i], trts[j], ],
                       seTE.random[trts[i], trts[j], ],
-                      method.tau = x$method.tau, method.tau.ci = "")
+                      method.tau = method.tau, method.tau.ci = "")
       #
       for (k in levs) {
         res.ijk <-
@@ -209,6 +239,8 @@ subgroup.netmeta <- function(x, subgroup, only.connected = FALSE,
                      k = A.matrix[trts[i], trts[j], k],
                      TE = TE.random[trts[i], trts[j], k],
                      seTE = seTE.random[trts[i], trts[j], k],
+                     lower = lower.random[trts[i], trts[j], k],
+                     upper = upper.random[trts[i], trts[j], k],
                      tau2 = tau2.matrix[trts[i], trts[j], k],
                      tau = tau.matrix[trts[i], trts[j], k],
                      Q = m.ij$Q, df.Q = m.ij$df.Q, pval.Q = m.ij$pval.Q)
@@ -243,11 +275,18 @@ subgroup.netmeta <- function(x, subgroup, only.connected = FALSE,
   x$common <- common
   x$random <- random
   x$method.tau <- method.tau
+  x$level.ma <- level.ma
+  x$backtransf <- backtransf
   x$nchar.trts <- nchar.trts
   #
   res <- list(common = res.c, random = res.r,
+              #
               TE.common = TE.common, seTE.common = seTE.common,
+              lower.common = lower.common, upper.common = upper.common,
+              #
               TE.random = TE.random, seTE.random = seTE.random,
+              lower.random = lower.random, upper.random = upper.random,
+              #
               tau2.matrix = tau2.matrix, tau.matrix = tau.matrix,
               A.matrix = A.matrix,
               networks = networks,
@@ -278,6 +317,7 @@ subgroup <- function(x, ...)
 print.subgroup.netmeta <- function(x,
                                    common = x$x$common,
                                    random = x$x$random,
+                                   backtransf = x$x$backtransf,
                                    #
                                    nchar.trts = x$x$nchar.trts,
                                    #
@@ -293,6 +333,7 @@ print.subgroup.netmeta <- function(x,
                                    zero.pval = gs("zero.pval"),
                                    JAMA.pval = gs("JAMA.pval"),
                                    #
+                                   print.se = !backtransf,
                                    print.tau2 = gs("print.tau2"),
                                    print.tau = gs("print.tau"),
                                    print.Q = gs("print.Q"),
@@ -309,6 +350,7 @@ print.subgroup.netmeta <- function(x,
   #
   chklogical(common)
   chklogical(random)
+  chklogical(backtransf)
   #
   if (is.null(nchar.trts))
     nchar.trts <- 666
@@ -327,6 +369,7 @@ print.subgroup.netmeta <- function(x,
   chklogical(zero.pval)
   chklogical(JAMA.pval)
   #
+  chklogical(print.se)
   chklogical(print.tau2)
   chklogical(print.tau)
   chklogical(print.Q)
@@ -339,20 +382,52 @@ print.subgroup.netmeta <- function(x,
   
   # Get rid of warning 'Undefined global functions or variables'
   .seTE <- .studlab <- .TE <- .treat1 <- .treat2 <- comparison <-
-    df.Q <- subnet <- treat1 <- treat2 <- NULL
+    df.Q <- subnet <- treat1 <- treat2 <- k <- TE <- NULL
   #
   trts <- x$x$trts
   trts.abbr <- treats(x$x$trts, nchar.trts)
+  sm <- sm.lab <- x$x$sm
+  #
+  if (!backtransf & (is_relative_effect(sm) | sm == "VE"))
+    sm.lab <- paste0("log", if (sm == "VE") "VR" else sm)
+  ##
+  ci.lab <- paste0(round(100 * x$x$level.ma, 1), "%-CI")
   
   if (common) {
     cat("Common effects model: \n\n")
     #
     dat.common <- x$common %>% filter(is.na(df.Q) | df.Q > 0)
     #
+    if (backtransf) {
+      dat.common$TE    <- backtransf(dat.common$TE, sm)
+      dat.common$lower <- backtransf(dat.common$lower, sm)
+      dat.common$upper <- backtransf(dat.common$upper, sm)
+      #
+      # Switch lower and upper limit for VE if results have been
+      # backtransformed
+      #
+      if (sm == "VE") {
+        tmp.l <- dat.common$lower
+        dat.common$lower <- dat.common$upper
+        dat.common$upper <- tmp.l
+      }
+    }
+    #
     dat.common$TE <-
       formatN(dat.common$TE, digits = digits, text.NA = ".")
-    dat.common$seTE <-
-      formatN(dat.common$seTE, digits = digits.se, text.NA = ".")
+    #
+    if (print.se)
+      dat.common$seTE <-
+        formatN(dat.common$seTE, digits = digits.se, text.NA = ".")
+    else
+      dat.common$seTE <- NULL
+    #
+    dat.common$ci <-
+      formatCI(formatN(round(dat.common$lower, digits),
+                       digits, ".", big.mark = big.mark),
+               formatN(round(dat.common$upper, digits),
+                       digits, ".", big.mark = big.mark))
+    #
     if (print.Q) {
       dat.common$Q <-
         formatN(dat.common$Q, digits = digits.Q, text.NA = ".")
@@ -360,7 +435,8 @@ print.subgroup.netmeta <- function(x,
         formatN(dat.common$df.Q, digits = 0, text.NA = ".")
       dat.common$pval.Q <-
         formatPT(dat.common$pval.Q,
-                 digits = digits.pval.Q, scientific = scientific.pval,
+                 digits = digits.pval.Q, lab.NA = ".",
+                 scientific = scientific.pval,
                  zero = zero.pval, JAMA = JAMA.pval)
     }
     #
@@ -371,10 +447,15 @@ print.subgroup.netmeta <- function(x,
         factor(dat.common$treat2, levels = trts, labels = trts.abbr)
     }
     #
-    prmatrix(dat.common %>% select(-treat1, -treat2),
-             quote = FALSE, right = TRUE,
-             rowlab =
-               paste(dat.common$treat1, dat.common$treat2, sep = x$x$sep.trts))
+    rowlabs <- paste(dat.common$treat1, dat.common$treat2, sep = x$x$sep.trts)
+    dat.common %<>% select(-treat1, -treat2) %>%
+      select(subgroup, k, TE, if (print.se) "seTE", ci,
+             if (print.Q) "Q", if (print.Q) "df.Q", if (print.Q) "pval.Q")
+    #
+    names(dat.common)[names(dat.common) == "TE"] <- sm.lab
+    names(dat.common)[names(dat.common) == "ci"] <- ci.lab
+    #
+    prmatrix(dat.common, quote = FALSE, right = TRUE, rowlab = rowlabs)
     #
     if (random)
       cat("\n")
@@ -387,12 +468,37 @@ print.subgroup.netmeta <- function(x,
     #
     dat.random <- x$random %>% filter(is.na(df.Q) | df.Q > 0)
     #
+    if (backtransf) {
+      dat.random$TE    <- backtransf(dat.random$TE, sm)
+      dat.random$lower <- backtransf(dat.random$lower, sm)
+      dat.random$upper <- backtransf(dat.random$upper, sm)
+      #
+      # Switch lower and upper limit for VE if results have been
+      # backtransformed
+      #
+      if (sm == "VE") {
+        tmp.l <- dat.random$lower
+        dat.random$lower <- dat.random$upper
+        dat.random$upper <- tmp.l
+      }
+    }
+    #
     dat.random$TE <-
       formatN(dat.random$TE, digits = digits, text.NA = ".",
               big.mark = big.mark)
-    dat.random$seTE <-
-      formatN(dat.random$seTE, digits = digits.se, text.NA = ".",
-              big.mark = big.mark)
+    #
+    if (print.se)
+      dat.random$seTE <-
+        formatN(dat.random$seTE, digits = digits.se, text.NA = ".")
+    else
+      dat.random$seTE <- NULL
+    #
+    dat.random$ci <-
+      formatCI(formatN(round(dat.random$lower, digits),
+                     digits, ".", big.mark = big.mark),
+             formatN(round(dat.random$upper, digits),
+                     digits, ".", big.mark = big.mark))
+    #
     if (print.Q) {
       dat.random$Q <-
         formatN(dat.random$Q, digits = digits.Q, text.NA = ".",
@@ -402,16 +508,19 @@ print.subgroup.netmeta <- function(x,
                 big.mark = big.mark)
       dat.random$pval.Q <-
         formatPT(dat.random$pval.Q,
-                 digits = digits.pval.Q, scientific = scientific.pval,
+                 digits = digits.pval.Q, lab.NA = ".",
+                 scientific = scientific.pval,
                  zero = zero.pval, JAMA = JAMA.pval)
     }
     #
     if (print.tau2)
       dat.random$tau2 <-
-      formatPT(dat.random$tau2, digits = digits.tau2, big.mark = big.mark)
+      formatPT(dat.random$tau2, digits = digits.tau2, lab.NA = ".",
+               big.mark = big.mark)
     if (print.tau2)
       dat.random$tau <-
-      formatPT(dat.random$tau, digits = digits.tau, big.mark = big.mark)
+      formatPT(dat.random$tau, digits = digits.tau, lab.NA = ".",
+               big.mark = big.mark)
     #
     if (any(trts != trts.abbr)) {
       dat.random$treat1 <-
@@ -420,10 +529,16 @@ print.subgroup.netmeta <- function(x,
         factor(dat.random$treat2, levels = trts, labels = trts.abbr)
     }
     #
-    prmatrix(dat.random %>% select(-treat1, -treat2),
-             quote = FALSE, right = TRUE,
-             rowlab =
-               paste(dat.random$treat1, dat.random$treat2, sep = x$x$sep.trts))
+    rowlabs <- paste(dat.random$treat1, dat.random$treat2, sep = x$x$sep.trts)
+    dat.random %<>% select(-treat1, -treat2) %>%
+      select(subgroup, k, TE, if (print.se) "seTE", ci,
+             if (print.tau2) "tau2", if (print.tau) "tau",
+             if (print.Q) "Q", if (print.Q) "df.Q", if (print.Q) "pval.Q")
+    #
+    names(dat.random)[names(dat.random) == "TE"] <- sm.lab
+    names(dat.random)[names(dat.random) == "ci"] <- ci.lab
+    #
+    prmatrix(dat.random, quote = FALSE, right = TRUE, rowlab = rowlabs)
   }
   #
   # Print details of network meta-analysis methods